moved mapdata integration test out of tests/poutine/
